refactor: Model Package내 클래스위치를 사용하는 UI 패키지로 이동
