--- app/src/main/java/com/kotlinisgood/boomerang/model/Video.kt
@@ -1,7 +1,0 @@
-package com.kotlinisgood.boomerang.model
-
-data class Video(
-    val uri: String,
-    val startSecond: Int,
-    val endSecond: Int,
-)
