+++ app/src/main/java/com/kotlinisgood/boomerang/ui/home/HomeViewModel.kt
@@ -5,7 +5,6 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.kotlinisgood.boomerang.database.entity.MediaMemo
-import com.kotlinisgood.boomerang.model.OrderState
 import com.kotlinisgood.boomerang.repository.AppRepository
 import com.kotlinisgood.boomerang.repository.SharedPrefDataSource
 import dagger.hilt.android.lifecycle.HiltViewModel
