+++ app/src/main/res/layout/item_rv_home_show_audio.xml
@@ -4,10 +4,10 @@
     xmlns:tools="http://schemas.android.com/tools">
 
     <data>
-        <import type="com.kotlinisgood.boomerang.model.OrderState"/>
+        <import type="com.kotlinisgood.boomerang.ui.home.OrderState"/>
         <variable
             name="orderState"
-            type="OrderState" />
+            type="com.kotlinisgood.boomerang.ui.home.OrderState" />
         <variable
             name="mediaMemo"
             type="com.kotlinisgood.boomerang.database.entity.MediaMemo" />
